+++ src/builtins/complete.rs
@@ -180,6 +180,7 @@ where
                     stdout,
                 );
             }
+            "enabled" => SHELL_BUILTINS,
             "helptopic" => crate::builtins::help::HELP_TOPICS,
             "keyword" => SHELL_KEYWORDS,
             "signal" => crate::builtins::trap::SIGNALS.as_slice(),